pyproject.toml
[project]
name = "deadman-switch"
version = "0.1.0"
description = "Deadman switch backend system with distributed task processing"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.18",
    "itsdangerous>=2.2.0",
    "werkzeug>=3.1.0",
]
